Removed dead code, small refactorings and cleanup of commands
