+++ discord/command_handlers.go
@@ -1,0 +1,377 @@
+package discord
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"github.com/denverquane/amongusdiscord/metrics"
+	"log"
+	"strconv"
+	"strings"
+
+	"github.com/bwmarrin/discordgo"
+	"github.com/denverquane/amongusdiscord/discord/command"
+	"github.com/denverquane/amongusdiscord/storage"
+	"github.com/nicksnyder/go-i18n/v2/i18n"
+)
+
+const MaxDebugMessageSize = 1980
+
+//TODO cache/preconstruct these (no reason to make them fresh everytime help is called, except for the prefix...)
+func ConstructEmbedForCommand(prefix string, cmd command.Command, sett *storage.GuildSettings) *discordgo.MessageEmbed {
+	return &discordgo.MessageEmbed{
+		URL:   "",
+		Type:  "",
+		Title: cmd.Emoji + " " + strings.Title(cmd.Command),
+		Description: sett.LocalizeMessage(cmd.Description,
+			map[string]interface{}{
+				"CommandPrefix": sett.CommandPrefix,
+			}),
+		Timestamp: "",
+		Color:     15844367, //GOLD
+		Image:     nil,
+		Thumbnail: nil,
+		Video:     nil,
+		Provider:  nil,
+		Author:    nil,
+		Fields: []*discordgo.MessageEmbedField{
+			{
+				Name: sett.LocalizeMessage(&i18n.Message{
+					ID:    "commands.ConstructEmbedForCommand.Fields.Example",
+					Other: "Example",
+				}),
+				Value:  "`" + fmt.Sprintf("%s %s", prefix, cmd.Example) + "`",
+				Inline: false,
+			},
+			{
+				Name: sett.LocalizeMessage(&i18n.Message{
+					ID:    "commands.ConstructEmbedForCommand.Fields.Arguments",
+					Other: "Arguments",
+				}),
+				Value:  "`" + sett.LocalizeMessage(cmd.Arguments) + "`",
+				Inline: false,
+			},
+			{
+				Name: sett.LocalizeMessage(&i18n.Message{
+					ID:    "commands.ConstructEmbedForCommand.Fields.Aliases",
+					Other: "Aliases",
+				}),
+				Value:  strings.Join(cmd.Aliases, ", "),
+				Inline: false,
+			},
+		},
+	}
+}
+
+func (bot *Bot) HandleCommand(isAdmin, isPermissioned bool, sett *storage.GuildSettings, s *discordgo.Session, g *discordgo.Guild, m *discordgo.MessageCreate, args []string) {
+	prefix := sett.CommandPrefix
+	cmd := command.GetCommand(args[0])
+
+	gsr := GameStateRequest{
+		GuildID:     m.GuildID,
+		TextChannel: m.ChannelID,
+	}
+
+	if cmd.CommandType != command.Null {
+		log.Print(fmt.Sprintf("\"%s\" command typed by User %s\n", cmd.Command, m.Author.ID))
+	}
+
+	//only allow admins to execute admin commands
+	if cmd.IsAdmin && !isAdmin {
+		s.ChannelMessageSend(m.ChannelID, sett.LocalizeMessage(&i18n.Message{
+			ID:    "message_handlers.handleMessageCreate.noPerms",
+			Other: "User does not have the required permissions to execute this command!",
+		}))
+	} else if cmd.IsOperator && (!isPermissioned && !isAdmin) {
+		s.ChannelMessageSend(m.ChannelID, sett.LocalizeMessage(&i18n.Message{
+			ID:    "message_handlers.handleMessageCreate.noPerms",
+			Other: "User does not have the required permissions to execute this command!",
+		}))
+	} else {
+		//broadly speaking, most commands issue at minimum 1 discord request, and delete a user's message.
+		//Very approximately, at least
+		bot.MetricsCollector.RecordDiscordRequests(bot.RedisInterface.client, metrics.MessageCreateDelete, 2)
+
+		switch cmd.CommandType {
+		case command.Help:
+			if len(args[1:]) == 0 {
+				embed := helpResponse(isAdmin, isPermissioned, prefix, command.AllCommands, sett)
+				s.ChannelMessageSendEmbed(m.ChannelID, &embed)
+			} else {
+				cmd = command.GetCommand(args[1])
+				if cmd.CommandType != command.Null {
+					embed := ConstructEmbedForCommand(prefix, cmd, sett)
+					s.ChannelMessageSendEmbed(m.ChannelID, embed)
+				} else {
+					s.ChannelMessageSend(m.ChannelID, sett.LocalizeMessage(&i18n.Message{
+						ID:    "commands.HandleCommand.Help.notFound",
+						Other: "I didn't recognize that command! View `help` for all available commands!",
+					}))
+				}
+			}
+			break
+
+		case command.New:
+			room, region := getRoomAndRegionFromArgs(args[1:], sett)
+
+			bot.handleNewGameMessage(s, m, g, sett, room, region)
+			break
+
+		case command.End:
+			log.Println("User typed end to end the current game")
+
+			dgs := bot.RedisInterface.GetReadOnlyDiscordGameState(gsr)
+			if v, ok := bot.EndGameChannels[dgs.ConnectCode]; ok {
+				v <- EndGameMessage{EndGameType: EndAndWipe}
+			}
+			delete(bot.EndGameChannels, dgs.ConnectCode)
+
+			bot.applyToAll(dgs, false, false)
+
+			break
+
+		case command.Pause:
+			lock, dgs := bot.RedisInterface.GetDiscordGameStateAndLock(gsr)
+			if lock == nil {
+				break
+			}
+			dgs.Running = !dgs.Running
+
+			bot.RedisInterface.SetDiscordGameState(dgs, lock)
+			if !dgs.Running {
+				bot.applyToAll(dgs, false, false)
+			}
+
+			edited := dgs.Edit(s, bot.gameStateResponse(dgs, sett))
+			if edited {
+				bot.MetricsCollector.RecordDiscordRequests(bot.RedisInterface.client, metrics.MessageEdit, 1)
+			}
+			break
+
+		case command.Refresh:
+			bot.RefreshGameStateMessage(gsr, sett)
+			break
+		case command.Link:
+			if len(args[1:]) < 2 {
+				embed := ConstructEmbedForCommand(prefix, cmd, sett)
+				s.ChannelMessageSendEmbed(m.ChannelID, embed)
+			} else {
+				lock, dgs := bot.RedisInterface.GetDiscordGameStateAndLock(gsr)
+				if lock == nil {
+					break
+				}
+				bot.linkPlayer(s, dgs, args[1:])
+				bot.RedisInterface.SetDiscordGameState(dgs, lock)
+
+				edited := dgs.Edit(s, bot.gameStateResponse(dgs, sett))
+				if edited {
+					bot.MetricsCollector.RecordDiscordRequests(bot.RedisInterface.client, metrics.MessageEdit, 1)
+				}
+			}
+			break
+
+		case command.Unlink:
+			if len(args[1:]) == 0 {
+				embed := ConstructEmbedForCommand(prefix, cmd, sett)
+				s.ChannelMessageSendEmbed(m.ChannelID, embed)
+			} else {
+
+				userID, err := extractUserIDFromMention(args[1])
+				if err != nil {
+					log.Println(err)
+				} else {
+					log.Print(fmt.Sprintf("Removing player %s", userID))
+					lock, dgs := bot.RedisInterface.GetDiscordGameStateAndLock(gsr)
+					if lock == nil {
+						break
+					}
+					dgs.ClearPlayerData(userID)
+
+					bot.RedisInterface.SetDiscordGameState(dgs, lock)
+
+					//update the state message to reflect the player leaving
+					edited := dgs.Edit(s, bot.gameStateResponse(dgs, sett))
+					if edited {
+						bot.MetricsCollector.RecordDiscordRequests(bot.RedisInterface.client, metrics.MessageEdit, 1)
+					}
+				}
+			}
+			break
+		case command.UnmuteAll:
+			dgs := bot.RedisInterface.GetReadOnlyDiscordGameState(gsr)
+			bot.applyToAll(dgs, false, false)
+			break
+
+		case command.Settings:
+			premStatus := bot.PostgresInterface.GetGuildPremiumStatus(m.GuildID)
+			bot.HandleSettingsCommand(s, m, sett, args, premStatus != 0)
+			break
+
+		case command.Cache:
+			if len(args[1:]) == 0 {
+				embed := ConstructEmbedForCommand(prefix, cmd, sett)
+				s.ChannelMessageSendEmbed(m.ChannelID, embed)
+			} else {
+				userID, err := extractUserIDFromMention(args[1])
+				if err != nil {
+					log.Println(err)
+					s.ChannelMessageSend(m.ChannelID, "I couldn't find a user by that name or ID!")
+					break
+				}
+				if len(args[2:]) == 0 {
+					cached := bot.RedisInterface.GetUsernameOrUserIDMappings(m.GuildID, userID)
+					if len(cached) == 0 {
+						s.ChannelMessageSend(m.ChannelID, sett.LocalizeMessage(&i18n.Message{
+							ID:    "commands.HandleCommand.Cache.emptyCachedNames",
+							Other: "I don't have any cached player names stored for that user!",
+						}))
+					} else {
+						buf := bytes.NewBuffer([]byte(sett.LocalizeMessage(&i18n.Message{
+							ID:    "commands.HandleCommand.Cache.cachedNames",
+							Other: "Cached in-game names:",
+						})))
+						buf.WriteString("\n```\n")
+						for n := range cached {
+							buf.WriteString(fmt.Sprintf("%s\n", n))
+						}
+						buf.WriteString("```")
+
+						s.ChannelMessageSend(m.ChannelID, buf.String())
+					}
+				} else if strings.ToLower(args[2]) == "clear" || strings.ToLower(args[2]) == "c" {
+					err := bot.RedisInterface.DeleteLinksByUserID(m.GuildID, userID)
+					if err != nil {
+						log.Println(err)
+					} else {
+						s.ChannelMessageSend(m.ChannelID, sett.LocalizeMessage(&i18n.Message{
+							ID:    "commands.HandleCommand.Cache.Success",
+							Other: "Successfully deleted all cached names for that user!",
+						}))
+					}
+				}
+			}
+			break
+
+		case command.Privacy:
+			if m.Author != nil {
+				var arg = ""
+				if len(args[1:]) > 0 {
+					arg = args[1]
+				}
+				if arg == "" || (arg != "showme" && arg != "optin" && arg != "optout") {
+					embed := ConstructEmbedForCommand(prefix, cmd, sett)
+					s.ChannelMessageSendEmbed(m.ChannelID, embed)
+				} else {
+					embed := bot.privacyResponse(m.GuildID, m.Author.ID, arg, sett)
+					s.ChannelMessageSendEmbed(m.ChannelID, embed)
+				}
+			}
+			break
+
+		case command.Info:
+			embed := bot.infoResponse(sett)
+			_, err := s.ChannelMessageSendEmbed(m.ChannelID, embed)
+			if err != nil {
+				log.Println(err)
+			}
+			break
+
+		case command.DebugState:
+			if m.Author != nil {
+				state := bot.RedisInterface.GetReadOnlyDiscordGameState(gsr)
+				if state != nil {
+					jBytes, err := json.MarshalIndent(state, "", "  ")
+					if err != nil {
+						log.Println(err)
+					} else {
+						for i := 0; i < len(jBytes); i += MaxDebugMessageSize {
+							end := i + MaxDebugMessageSize
+							if end > len(jBytes) {
+								end = len(jBytes)
+							}
+							s.ChannelMessageSend(m.ChannelID, fmt.Sprintf("```JSON\n%s\n```", jBytes[i:end]))
+						}
+					}
+				}
+			}
+			break
+
+		case command.Ascii:
+			if len(args[1:]) == 0 {
+				s.ChannelMessageSend(m.ChannelID, AsciiCrewmate)
+			} else {
+				id, err := extractUserIDFromMention(args[1])
+				if id == "" || err != nil {
+					s.ChannelMessageSend(m.ChannelID, "I couldn't find a user by that name or ID!")
+				} else {
+					imposter := false
+					count := 1
+					if len(args[2:]) > 0 {
+						if args[2] == "true" || args[2] == "t" {
+							imposter = true
+						}
+						if len(args[3:]) > 0 {
+							if itCount, err := strconv.Atoi(args[3]); err == nil {
+								count = itCount
+							}
+						}
+					}
+					s.ChannelMessageSend(m.ChannelID, AsciiStarfield(sett, args[1], imposter, count))
+				}
+			}
+			break
+		case command.Stats:
+			premStatus := bot.PostgresInterface.GetGuildPremiumStatus(m.GuildID)
+			if len(args[1:]) == 0 {
+				embed := ConstructEmbedForCommand(prefix, cmd, sett)
+				s.ChannelMessageSendEmbed(m.ChannelID, embed)
+			} else {
+				userID, err := extractUserIDFromMention(args[1])
+				if userID == "" || err != nil {
+					if strings.ReplaceAll(strings.ToLower(args[1]), "\"", "") == "guild" {
+						_, err := s.ChannelMessageSendEmbed(m.ChannelID, bot.GuildStatsEmbed(m.GuildID, sett, premStatus))
+						if err != nil {
+							log.Println(err)
+						}
+					} else {
+						s.ChannelMessageSend(m.ChannelID, "I didn't recognize that user, or you mistyped 'guild'...")
+					}
+
+				} else {
+					s.ChannelMessageSendEmbed(m.ChannelID, bot.UserStatsEmbed(userID, m.GuildID, sett, premStatus))
+				}
+
+			}
+			break
+		case command.Premium:
+			premStatus := bot.PostgresInterface.GetGuildPremiumStatus(m.GuildID)
+			if len(args[1:]) == 0 {
+				s.ChannelMessageSendEmbed(m.ChannelID, premiumEmbedResponse(premStatus, sett))
+			} else {
+				arg := strings.ToLower(args[1])
+				if isAdmin {
+					if arg == "invite" || arg == "inv" {
+						_, err := s.ChannelMessageSendEmbed(m.ChannelID, premiumInvitesEmbed(premStatus, sett))
+						if err != nil {
+							log.Println(err)
+						}
+					}
+				} else {
+					s.ChannelMessageSend(m.ChannelID, "Viewing the premium invites is an Admin command")
+				}
+			}
+			break
+		default:
+			s.ChannelMessageSend(m.ChannelID, sett.LocalizeMessage(&i18n.Message{
+				ID:    "commands.HandleCommand.default",
+				Other: "Sorry, I didn't understand that command! Please see `{{.CommandPrefix}} help` for commands",
+			},
+				map[string]interface{}{
+					"CommandPrefix": prefix,
+				}))
+			break
+		}
+	}
+
+	deleteMessage(s, m.ChannelID, m.Message.ID)
+}
