+++ discord/healthcheck.go
@@ -33,16 +33,6 @@ func StartHealthCheckServer(port string) {
 			w.WriteHeader(http.StatusTooManyRequests)
 			log.Println("I'M BEING RATE-LIMITED BY DISCORD")
 			return
-			//orgId := os.Getenv("SCW_ORGANIZATION_ID")
-			//accessKey := os.Getenv("SCW_ACCESS_KEY")
-			//secretKey := os.Getenv("SCW_SECRET_KEY")
-			//nodeID := os.Getenv("SCW_NODE_ID")
-			//if orgId == "" || accessKey == "" || secretKey == "" || nodeID == "" {
-			//	log.Println("One of the Scaleway credentials was null, not replacing any nodes!")
-			//	return
-			//}
-			//
-			//metrics.TerminateScalewayNode(orgId, accessKey, secretKey, nodeID)
 		}
 	})
